pyproject.toml
[tool.poetry]
name = "libms"
version = "0.1.0"
description = ""
authors = ["Nkoroi <nkoroi.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14.0"
django-pwa = "^1.1.0"
django-cors-headers = "^3.14.0"
django-import-export = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
